Remove `TODO: Should uninitialized contain peers?`
My decision is that this is fine to keep. I don't think an `uninitialized`
server needs any fields, but if we were to remove all the state in the server
struct and move it into the `ServerMode` structs, that would result in a lot
of code duplication for not a lot of gain.